tests/requirements.txt
pytest
pytest-env
pytest-cov
pytest-django
psycopg2-binary
django-environ
